refactor: language, address 테이블 삭제

+++ server/src/main/java/com/codueon/boostUp/domain/lesson/dto/PostLesson.java
@@ -10,10 +10,10 @@ import java.util.List;
 @NoArgsConstructor
 public class PostLesson {
         private String title;
-        private List<Long> languages;
+        private List<Integer> languages;
         private String company;
         private Integer career;
-        private List<Long> address;
+        private List<Integer> address;
         private Integer cost;
         private String introduction;
         private String detailCompany;
@@ -25,10 +25,10 @@ public class PostLesson {
         @Builder
         public PostLesson(
                           String title,
-                          List<Long> languages,
+                          List<Integer> languages,
                           String company,
                           Integer career,
-                          List<Long> address,
+                          List<Integer> address,
                           Integer cost,
                           String introduction,
                           String detailCompany,
